Added VS2017 Windows Application - Flash Programmer
This is still not finished but does work to a point. It does miss off the last parts of the GetID function information returned from the Arduino though. Writing, reading and erasing do work but I would use with caution and test everything before trying out on a Flash chip you can't do without!!
Later features will be to correct the GetID function, streaming the read data and allowing a sector erase as well as the entire chip erase function.
I'll leave in the old commands so terminal use will still work.
Enjoy!